+++ lib/karafka/web/config.rb
@@ -6,6 +6,11 @@ module Karafka
     class Config
       include ::Karafka::Core::Configurable
 
+      #setting(:secret_key_base, default: false, constructor: ->(default) {
+      #    ::Object.const_defined?(:Rails) ? Rails.application.secrets[:secret_key_base] : default
+      #  }
+      #)
+
       # How long do we consider the process alive without receiving status info about it
       # For this long we also display dead processes (shutdown) in the UI
       # This is used both in the processing for eviction and in the UI
@@ -80,9 +85,25 @@ module Karafka
       end
 
       setting :ui do
-        # Should the payload be decrypted for the Pro Web UI. Default to `false` due to security
-        # reasons
-        setting :decrypt, default: false
+        setting :explorer do
+          # On a per topic view it is expensive (one call per partition) to get offsets and
+          # manage the states for aggregated view.
+          #
+          # This is the max partitions we query for and if there are more, their data will not be
+          # displayed and we will show a warning. This prevents the system from being overloaded
+          # as for the negative lookups we need to perform additionals calls to Kafka
+          #
+          # If the system is responsive for you with this number, you can increase it to match
+          # your cluster setup
+          #
+          # librdkafka currently does not support batch watermark offsets aggregations and this is
+          # why this limit is introduced.
+          setting :max_aggregable_partitions, default: 50
+
+          # Should the payload be decrypted for the Pro Web UI. Default to `false` due to security
+          # reasons
+          setting :decrypt, default: false
+        end
 
         # How many elements should we display on pages that support pagination
         setting :per_page, default: 25