+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 # Karafka Web changelog
 
 ## 0.7.0 (Unreleased)
+- **[Feature]** Introduce per-topic data exploration (limited up to 50 first partitions by default).
 - [Improvement] Switch to offset based pagination instead of per-page pagination.
 - [Improvement] Avoid double-reading of watermark offsets for explorer and errors display.
 - [Improvement] When no params needed for a page, do not include empty params.
 - [Improvement] Do not include page when page is 1.
+- [Refactor] Replace `ui.decrypt` config option with `ui.explorer.decrypt`
 - [Refactor] Reorganize pagination engine to support offset based pagination.
 
 ## 0.6.1 (2023-06-25)